remove: aeonik font artifacts
